Move: 로그아웃 버튼 파일이름 변경 및 이동 #55
