+++ apps/front-end/src/components/ui/log-out-button.tsx
